Move inline tests to tests/ directory
Extract #[cfg(test)] blocks from types.rs, commands.rs, assets.rs
into separate integration test files under tests/.
